Prover and verifier split onto their own submodules
